feat(firefox): implement frame.goto / frame.waitForNavigation
Some corner cases regarding iframes being detached during navigation
are not yet supported.
